tarifa calculo estimado. Ruta obtencion de parametros para obtener el calculo de tarifa, Solicitud comunicacion con las apis. Detalles en cada servicio

+++ solicitud-service/src/main/java/com/tpibakend/solicitud_service/infraestructure/controller/dto/RequestCreateRequest.java
@@ -9,11 +9,11 @@ public record RequestCreateRequest(
         String clientPhone,
         BigDecimal containerWeight,
         BigDecimal containerVolume,
-        BigDecimal originLat,
-        BigDecimal originLng,
+        Double originLat,
+        Double originLng,
         String originAddress,
-        BigDecimal destinationLat,
-        BigDecimal destinationLng,
+        Double destinationLat,
+        Double destinationLng,
         String destinationAddress
 ) {
 }
